feat: Correction in some migrations.

+++ backend/src/server/database/migrations/0004_create_billings.ts
@@ -11,12 +11,12 @@ export async function up(knex: Knex) {
         .inTable(ETableNames.client).onUpdate('CASCADE').onDelete('RESTRICT');
       table.date('dateReference').notNullable();
       table.decimal('amount', 9,2 ).notNullable();
-      table.integer('invoiceNumber').notNullable();
+      table.bigInteger('invoiceNumber').notNullable();
       table.date('invoiceDate').notNullable();
       table.date('dueDate').notNullable();
       table.decimal('amountPaid', 9,2).notNullable();
       table.date('paymentDate').notNullable();
-      table.integer('status').notNullable();
+      table.bigInteger('status').notNullable();
 
       table.comment('Table to insert "bills" in billing table ');
     })
