+++ backend/src/server/database/migrations/0008_create_attendance.ts
@@ -7,14 +7,14 @@ export async function up(knex: Knex) {
     .schema
     .createTable(ETableNames.attendance, table => {
       table.bigIncrements('id').primary().index();
-      table.integer('clientId').index().notNullable().references('id')
+      table.bigInteger('clientId').index().notNullable().references('id')
         .inTable(ETableNames.client).onUpdate('CASCADE').onDelete('RESTRICT');
       table.date('dateReference').notNullable();
       table.string('issue').notNullable();
-      table.integer('userId').notNullable().references('id')
+      table.bigInteger('userId').notNullable().references('id')
         .inTable(ETableNames.user).onUpdate('CASCADE').onDelete('RESTRICT');
       table.date('followUp').notNullable();
-      table.integer('status').notNullable().references('id')
+      table.bigInteger('status').notNullable().references('id')
         .inTable(ETableNames.status).onUpdate('CASCADE').onDelete('RESTRICT');
 
       table.comment('Table to insert "attendance" in attendance table ');
